+++ tests/data/ex_prefer_no_split_subscripts.py
@@ -2,6 +2,9 @@ x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"] = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
 
 x[x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"] or "xxxxxxx"] = "xxxxxxxxxxxxxxxxxxxxxxxxx"
 
+if xxxxxxxxxxxxxxxxxxxxxxxxxx in ["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"]:
+    pass
+
 # output
 
 x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"] = (
@@ -11,3 +14,8 @@ x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"] = (
 x[x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"] or "xxxxxxx"] = (
     "xxxxxxxxxxxxxxxxxxxxxxxxx"
 )
+
+if xxxxxxxxxxxxxxxxxxxxxxxxxx in [
+    "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+]:
+    pass